feat: update architecture of authorization
